% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearErrors.r
\name{clearError}
\alias{clearError}
\title{clear errors from subroutine BDAT20}
\usage{
clearError(data)
}
\arguments{
\item{data}{result of subroutine BDAT20}
}
\value{
a corrected, i.e. NA'd version of the input
}
\description{
Returns the NA'd data set
}
\keyword{internal}
