% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errormessage.r
\name{errormessage}
\alias{errormessage}
\title{Return dedicated error message}
\usage{
errormessage(code)
}
\arguments{
\item{code}{error codes from subroutine BDAT20 (parameter 'Ifeh')}
}
\value{
returns the error message of the respective codes
}
\description{
Returns the correct error message of SUBROUTINE BDAT20 according to help
file of BDAT
}
\keyword{internal}
