% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformBDAT20.R
\name{transformBDAT20}
\alias{transformBDAT20}
\title{transform BDAT20-matrix}
\usage{
transformBDAT20(assort, value)
}
\arguments{
\item{assort}{a matrix, produced by calling Fortran-BDAT20-subroutine and
extracting one list element to a matrix}

\item{value}{character vector indicating return type: either "Vol", "Skl"
or "Fix"}
}
\value{
a transformed BDAT20-result, usually a data.frame
}
\description{
transforms an intermediate BDAT20-matrix into a data.frame,
names it appropriately and removes <zero>-entries.
}
\details{
Fortran subroutine BDAT20 returns a list with many entries for each
tree; most relevant are 'Vol', 'FixLng' and maybe 'Skl'. These elements are
transformed into a matrix and properly whiped into shape.
}
\keyword{internal}
